refactor: move control message mods to control mod
We group the control messages under a control mod, parallel to the
data messages.
