+++ src/cache/messages/control/delete_cache.rs
